Backend clean up, Controllers have applicable flash messages for user, relations updated, Menu Item development started

+++ app/Http/Controllers/TaskController.php
@@ -30,21 +30,34 @@ class TaskController extends Controller
         $data['user_id'] = auth()->id();
 
         Task::create($data);
+
         return redirect()->back()->with('flash', [
             'type' => 'success',
             'message' => 'Task created successfully.']);
     }
 
-    public function show(Task $task): Task
+    public function show(Task $task): Task|RedirectResponse
     {
+        if ($task->user_id !== auth()->id()) {
+            return redirect()->back()->with('flash', [
+                'type' => 'error',
+                'message' => 'You are not authorized to view this task.',
+            ]);
+        }
         return $task;
     }
 
     public function update(Request $request, Task $task): RedirectResponse
     {
+        if ($task->user_id !== auth()->id()) {
+            return redirect()->back()->with('flash', [
+                'type' => 'error',
+                'message' => 'You are not authorized to edit this task.',
+            ]);
+        }
+
         $data = $request->validate([
             'title' => 'required',
-            'user_id' => 'required', 'exists:users',
             'status' => ['nullable', new Enum(TaskStatus::class)],
             'due_date' => 'nullable|date'
         ]);
@@ -58,6 +71,13 @@ class TaskController extends Controller
 
     public function destroy(Task $task): RedirectResponse
     {
+        if ($task->user_id !== auth()->id()) {
+            return redirect()->back()->with('flash', [
+                'type' => 'error',
+                'message' => 'You are not authorized to delete this task.',
+            ]);
+        }
+
         $task->delete();
 
         return redirect()->back()->with('flash', [
