+++ app/Http/Controllers/VenueController.php
@@ -10,7 +10,7 @@ class VenueController extends Controller
 {
     public function index()
     {
-        $venues = Venue::whereHas('venueCoordinators', function ($query) {
+        $venues = Venue::whereHas('users', function ($query) {
             $query->where('user_id', auth()->id());
         })->with('venueMenuItems')->get();
         return Inertia::render('admin/venue-manager', [
@@ -31,21 +31,34 @@ class VenueController extends Controller
 
         $venue = Venue::create($data);
 
-        $venue->venueCoordinators()->create(['user_id' => auth()->id()]);
+        $venue->users()->attach(auth()->id());
 
-        return redirect()->back()->with('success', 'Venue created successfully.');
+        return redirect()->back()->with('flash', [
+            'type' => 'success',
+            'message' => 'Venue created successfully.'
+            ]);
     }
 
     public function show(Venue $venue)
     {
-        if (!auth()->user()->venues->contains($venue->id)) {
-            abort(403);
+        if (!$venue->users()->where('user_id', auth()->id())->exists()) {
+            return redirect()->back()->with('flash', [
+                'type' => 'error',
+                'message' => 'You are not authorized to view this venue.',
+            ]);
         }
         return $venue;
     }
 
     public function update(Request $request, Venue $venue)
     {
+        if (!$venue->users()->where('user_id', auth()->id())->exists()) {
+            return redirect()->back()->with('flash', [
+                'type' => 'error',
+                'message' => 'You are not authorized to edit this venue.',
+            ]);
+        }
+
         $data = $request->validate([
             'name' => ['required'],
             'minimumTableAmount' => ['required', 'integer'],
@@ -56,11 +69,21 @@ class VenueController extends Controller
 
         $venue->update($data);
 
-        return $venue;
+        return redirect()->back()->with('flash', [
+            'type' => 'success',
+            'message' => 'Venue updated successfully.'
+        ]);
     }
 
     public function destroy(Venue $venue)
     {
+        if (!$venue->users()->where('user_id', auth()->id())->exists()) {
+            return redirect()->back()->with('flash', [
+                'type' => 'error',
+                'message' => 'You are not authorized to delete this venue.',
+            ]);
+        }
+
         $venue->delete();
 
         return response()->json();
